Add files via upload
with this program I can communicate with many computers, using the same router. I need the port number of the others and their Ip address. I can open many windows to chat with different people. My port Number is in My message class and my Ip the program give it to me.